Reorganize files to create models folder
