+++ FSJ_django20_project/FSJ/forms.py
@@ -1,6 +1,6 @@
 from django.forms import *
 from django.utils.translation import gettext_lazy as _
-from .models_FSJUser import FSJUser
+from .models import FSJUser
 # Split the forms into packages to allow easier maintenance and multi-user modifications with less conflicts
 # Import the forms all into the standard forms.py page so all forms can exist in one place
 from .forms_student import *
