+++ FSJ_django20_project/FSJ/forms_adjudicator.py
@@ -1,4 +1,4 @@
-from .models_adjudicator import Adjudicator
+from .models import Adjudicator
 from .forms_modelform import ModelForm
 
 class AdjudicatorForm(ModelForm):
