+++ FSJ_django20_project/FSJ/forms_program.py
@@ -1,4 +1,4 @@
-from .models_program import Program
+from .models import Program
 from .forms_modelform import ModelForm
 
 #Modelform for a program, this restricts what fields will be enabled/disabled as well as widgets, etc
