+++ FSJ_django20_project/FSJ/forms_yearofstudy.py
@@ -1,4 +1,4 @@
-from .models_yearofstudy import YearOfStudy
+from .models import YearOfStudy
 from django.forms import CheckboxSelectMultiple
 from .forms_modelform import ModelForm
