+++ FSJ_django20_project/FSJ/models/__init__.py
@@ -1,0 +1,10 @@
+from .models_student import Student
+from .models_adjudicator import Adjudicator
+from .models_coordinator import Coordinator
+from .models_award import Award
+from .models_committee import Committee
+from .models_program import Program
+from .models_yearofstudy import YearOfStudy
+from .models_application import Application
+from .models_committee import Committee
+from .models_FSJUser import FSJUser
