+++ FSJ_django20_project/FSJ/models/models_yearofstudy.py
